pyproject.toml
[project]
name = "rwb-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "catboost>=1.2.10",
    "einops>=0.8.2",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.92",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
    "pillow>=12.2.0",
    "requests>=2.33.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.4.0",
    "streamlit>=1.56.0",
    "timm>=1.0.26",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]
